chore: move functions to proper files
