+++ pkg/provider/azure_standard.go
@@ -21,16 +21,12 @@ import (
 	"errors"
 	"fmt"
 	"hash/crc32"
-	"net"
 	"regexp"
 	"strconv"
 	"strings"
 	"sync"
 	"sync/atomic"
 	"time"
-	"unicode"
-
-	"sigs.k8s.io/cloud-provider-azure/pkg/consts"
 
 	"github.com/Azure/azure-sdk-for-go/services/compute/mgmt/2022-08-01/compute"
 	"github.com/Azure/azure-sdk-for-go/services/network/mgmt/2022-07-01/network"
@@ -44,6 +40,7 @@ import (
 	"k8s.io/utils/pointer"
 
 	azcache "sigs.k8s.io/cloud-provider-azure/pkg/cache"
+	"sigs.k8s.io/cloud-provider-azure/pkg/consts"
 	"sigs.k8s.io/cloud-provider-azure/pkg/metrics"
 )
 
@@ -376,112 +373,6 @@ func publicIPOwnsFrontendIP(service *v1.Service, fip *network.FrontendIPConfigur
 		}
 	}
 	return false
-}
-
-// There are two cases when a service owns the frontend IP config:
-// 1. The primary service, which means the frontend IP config is created after the creation of the service.
-// This means the name of the config can be tracked by the service UID.
-// 2. The secondary services must have their loadBalancer IP set if they want to share the same config as the primary
-// service. Hence, it can be tracked by the loadBalancer IP.
-// If the IP version is not empty, which means it is the secondary Service, it returns IP version of the Service FIP.
-func (az *Cloud) serviceOwnsFrontendIP(fip network.FrontendIPConfiguration, service *v1.Service) (bool, bool, network.IPVersion) {
-	var isPrimaryService bool
-	baseName := az.GetLoadBalancerName(context.TODO(), "", service)
-	if strings.HasPrefix(pointer.StringDeref(fip.Name, ""), baseName) {
-		klog.V(6).Infof("serviceOwnsFrontendIP: found primary service %s of the frontend IP config %s", service.Name, *fip.Name)
-		isPrimaryService = true
-		return true, isPrimaryService, ""
-	}
-
-	loadBalancerIPs := getServiceLoadBalancerIPs(service)
-	pipResourceGroup := az.getPublicIPAddressResourceGroup(service)
-	var pipNames []string
-	if len(loadBalancerIPs) == 0 {
-		if !requiresInternalLoadBalancer(service) {
-			pipNames = getServicePIPNames(service)
-			for _, pipName := range pipNames {
-				if pipName != "" {
-					pip, err := az.findMatchedPIP("", pipName, pipResourceGroup)
-					if err != nil {
-						klog.Warningf("serviceOwnsFrontendIP: unexpected error when finding match public IP of the service %s with name %s: %v", service.Name, pipName, err)
-						return false, isPrimaryService, ""
-					}
-					if publicIPOwnsFrontendIP(service, &fip, pip) {
-						return true, isPrimaryService, pip.PublicIPAddressPropertiesFormat.PublicIPAddressVersion
-					}
-				}
-			}
-		}
-		// it is a must that the secondary services set the loadBalancer IP or pip name
-		return false, isPrimaryService, ""
-	}
-
-	// for external secondary service the public IP address should be checked
-	if !requiresInternalLoadBalancer(service) {
-		for _, loadBalancerIP := range loadBalancerIPs {
-			pip, err := az.findMatchedPIP(loadBalancerIP, "", pipResourceGroup)
-			if err != nil {
-				klog.Warningf("serviceOwnsFrontendIP: unexpected error when finding match public IP of the service %s with loadBalancerIP %s: %v", service.Name, loadBalancerIP, err)
-				return false, isPrimaryService, ""
-			}
-
-			if publicIPOwnsFrontendIP(service, &fip, pip) {
-				return true, isPrimaryService, pip.PublicIPAddressPropertiesFormat.PublicIPAddressVersion
-			}
-			klog.V(6).Infof("serviceOwnsFrontendIP: the public IP with ID %s is being referenced by other service with public IP address %s "+
-				"OR it is of incorrect IP version", *pip.ID, *pip.IPAddress)
-		}
-
-		return false, isPrimaryService, ""
-	}
-
-	// for internal secondary service the private IP address on the frontend IP config should be checked
-	if fip.PrivateIPAddress == nil {
-		return false, isPrimaryService, ""
-	}
-	privateIPAddrVersion := network.IPv4
-	if net.ParseIP(*fip.PrivateIPAddress).To4() == nil {
-		privateIPAddrVersion = network.IPv6
-	}
-
-	privateIPEquals := false
-	for _, loadBalancerIP := range loadBalancerIPs {
-		if strings.EqualFold(*fip.PrivateIPAddress, loadBalancerIP) {
-			privateIPEquals = true
-			break
-		}
-	}
-	return privateIPEquals, isPrimaryService, privateIPAddrVersion
-}
-
-func (az *Cloud) getFrontendIPConfigNames(service *v1.Service) map[bool]string {
-	isDualStack := isServiceDualStack(service)
-	defaultLBFrontendIPConfigName := az.getDefaultFrontendIPConfigName(service)
-	return map[bool]string{
-		consts.IPVersionIPv4: getResourceByIPFamily(defaultLBFrontendIPConfigName, isDualStack, consts.IPVersionIPv4),
-		consts.IPVersionIPv6: getResourceByIPFamily(defaultLBFrontendIPConfigName, isDualStack, consts.IPVersionIPv6),
-	}
-}
-
-func (az *Cloud) getDefaultFrontendIPConfigName(service *v1.Service) string {
-	baseName := az.GetLoadBalancerName(context.TODO(), "", service)
-	subnetName := getInternalSubnet(service)
-	if subnetName != nil {
-		ipcName := fmt.Sprintf("%s-%s", baseName, *subnetName)
-
-		// Azure lb front end configuration name must not exceed 80 characters
-		maxLength := consts.FrontendIPConfigNameMaxLength - consts.IPFamilySuffixLength
-		if len(ipcName) > maxLength {
-			ipcName = ipcName[:maxLength]
-			// Cutting the string may result in char like "-" as the string end.
-			// If the last char is not a letter or '_', replace it with "_".
-			if !unicode.IsLetter(rune(ipcName[len(ipcName)-1:][0])) && ipcName[len(ipcName)-1:] != "_" {
-				ipcName = ipcName[:len(ipcName)-1] + "_"
-			}
-		}
-		return ipcName
-	}
-	return baseName
 }
 
 // This returns the next available rule priority level for a given set of security rules.
